Add rest of bepinex installer plugin code
... And a small edit to the PackageInstaller trait install_package and
uninstall_package functions to accept mutable self references, not
immutable.

In theory this is fine but it opens up problems if we plan on
multithreading file ops due to rust's mutable borrow rules. If this
becomes a problem (and we NEED mutability) we can rely on something like
a RefCell<T: TrackedFile> at the cost of some runtime peformance.